Tool: picked out MakeRoomScriptHeader() into separate unit
